Refactor method on automaton
