+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/AbstractBlackBoxModel.java
@@ -16,8 +16,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
-import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
-
 /**
  * @author Luma Zamarreño <zamarrenolm at aia.es>
  */
@@ -68,14 +66,5 @@ public abstract class AbstractBlackBoxModel implements BlackBoxModel, DynamicMod
     public List<BlackBoxModel> getModelsConnectedTo(DynaWaltzContext dynaWaltzContext) {
         // Default models are only connected to
         return Collections.emptyList();
-    }
-
-    protected void writeAutomatonBlackBoxModel(XMLStreamWriter writer, DynaWaltzContext context) throws XMLStreamException {
-        // Write the blackBoxModel object
-        writer.writeEmptyElement(DYN_URI, "blackBoxModel");
-        writer.writeAttribute("id", getDynamicModelId());
-        writer.writeAttribute("lib", getLib());
-        writer.writeAttribute("parFile", context.getParFile());
-        writer.writeAttribute("parId", getParameterSetId());
     }
 }
