+++ dynawaltz/src/main/java/com/powsybl/dynawaltz/dynamicmodels/staticref/AbstractBlackBoxModelWithStaticRef.java
@@ -8,15 +8,10 @@ package com.powsybl.dynawaltz.dynamicmodels.staticref;
 
 import com.powsybl.dynawaltz.DynaWaltzContext;
 import com.powsybl.dynawaltz.dynamicmodels.AbstractBlackBoxModelWithStaticId;
-import com.powsybl.dynawaltz.dynamicmodels.BlackBoxModel;
-import com.powsybl.dynawaltz.dynamicmodels.MacroConnector;
 import com.powsybl.dynawaltz.xml.MacroStaticReference;
-import org.apache.commons.lang3.tuple.Pair;
 
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamWriter;
-
-import java.util.List;
 
 import static com.powsybl.dynawaltz.xml.DynaWaltzXmlConstants.DYN_URI;
